prototype/requirements.txt
# Requirements for HELO Django Project
Django==5.0.2
channels==4.0.0
channels-redis==4.2.0
django-allauth==0.61.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-bootstrap5==23.4
django-notifications-hq==1.8.2
django-phonenumber-field[phonenumbers]==7.3.0
django-environ==0.11.2
python-dotenv==1.0.1
pillow==10.2.0
redis==5.0.1

# Optional: For production
# gunicorn
# psycopg2-binary

# Python version constraint
# python>=3.8,<3.12  # Uncomment this line if using pip-tools